fix: Fix race condition between dev build watchers
Running all build watchers in parallel will introduce race conditions as they depend on each other.
They cannot depend on each other directly in turborepo as the watchers don't indicate when they're done.
To get around this, we use a waiting command to wait for dependent builds to finish.

Note: This is not necessary for apps as nothing depends on these.